+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 ## Pending
 
 ### Added
+
 - Tested with Django 3.2. Only define ``default_app_config`` when using
   a version of Django earlier than 3.2.
 
 ### Fixed
+
 - Exclude python library paths from backtraces.
   ([PR #514](https://github.com/scoutapp/scout_apm_python/issues/514))
+- Preserve tracked request across asyncio tasks.
+  ([PR #469](https://github.com/scoutapp/scout_apm_python/issues/469))
 
 ## [2.18.0] 2021-02-09